move helper methods to protocols/horizon
